Added a "help" command to Game.java + help file
Help command is functional. Without additional arguments, the command
will call a voice from the sky (Really text from a text file) that will
list the commands. With arguments, it will list the syntax and use of
commands. Also added sassiness in the help command. If necessary,
sassiness may easily be removed. It probably should be removed.